+++ docs/.vitepress/components/pestanias/basico.vue
@@ -1,0 +1,3 @@
+<template>
+  <SisdaiPestanias></SisdaiPestanias>
+</template>