+++ docs/.vitepress/theme/NavegacionPrincipal.vue
@@ -1,11 +1,12 @@
 <script setup>
 import { ref } from 'vue'
 import { useData } from 'vitepress'
-import { isActive } from 'vitepress/dist/client/shared'
+import pkg from '../../../package.json'
 
 // https://vitepress.dev/reference/runtime-api#usedata
 const { theme, page } = useData()
 const navegacionPrincipal = ref(null)
+const cdn = import.meta.env.VITE_CDN_ARCHIVOS
 </script>
 
 <template>
@@ -14,6 +15,7 @@ const navegacionPrincipal = ref(null)
     ref="navegacionPrincipal"
   >
     <template #complementario>
+      <div class="nav-menu-contenedor">
       <a
         class="nav-hipervinculo"
         href="https://sisdai.conahcyt.mx"
@@ -22,6 +24,19 @@ const navegacionPrincipal = ref(null)
       >
         <b>IR A SISDAI</b>
       </a>
+      <a
+          class="nav-hipervinculo"
+          :href="pkg.repository.url"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          <img
+            class="nav-logo"
+            :src="`${cdn}gitlab-logo-500.png`"
+            alt="Repositorio de código sisdai-mapas"
+          /><b> {{ `v${pkg.version}` }} </b>
+        </a>
+    </div>
     </template>
     <ul class="nav-menu">
       <li
@@ -31,11 +46,9 @@ const navegacionPrincipal = ref(null)
         <a
           class="nav-hipervinculo"
           :class="{
-            'router-link-exact-active router-link-active': isActive(
-              page.relativePath,
-              nav.activeMatch || nav.link,
-              !!nav.activeMatch
-            ),
+            'router-link-exact-active router-link-active':
+              nav.activeMatch ==
+              page.relativePath.replace('index.md', '').split('/')[0],
           }"
           :href="nav.link"
           :target="nav.target"
@@ -51,5 +64,34 @@ const navegacionPrincipal = ref(null)
         </a>
       </li>
     </ul>
+
+    <!-- <ul class="nav-menu">
+      <li
+        v-for="nav in theme.nav"
+        :key="nav.text"
+      >
+        <a
+          class="nav-hipervinculo"
+          :class="{
+            'router-link-exact-active router-link-active': isActive(
+              page.relativePath,
+              nav.activeMatch || nav.link,
+              //!!nav.activeMatch
+            ),
+          }"
+          :href="nav.link"
+          :target="nav.target"
+          :rel="nav.rel"
+        >
+          <img
+            v-if="nav.img"
+            class="nav-logo"
+            :src="nav.img"
+            alt=""
+          />
+          {{ nav.text }}
+        </a>
+      </li>
+    </ul> -->
   </SisdaiNavegacionPrincipal>
 </template>