+++ src/componentes/index.js
@@ -13,9 +13,15 @@
 //   You should have received a copy of the GNU Lesser General Public License along
 //   with sisdai-componentes. If not, see <https://www.gnu.org/licenses/>.
 
+export { default as SisdaiAreaTexto } from './area-texto'
 export { default as SisdaiAudio } from './audio'
 export { default as SisdaiBotonFlotante } from './boton-flotante'
+export { default as SisdaiBotonRadio } from './boton-radio'
+export { default as SisdaiBotonesRadioGrupo } from './boton-radio-grupo'
+export { default as SisdaiCampoBase } from './campo-base'
 export { default as SisdaiCampoBusqueda } from './campo-busqueda'
+export { default as SisdaiCasillaVerificacion } from './casilla-verificacion'
+export { default as SisdaiCasillasVerificacionGrupo } from './casilla-verificacion-grupo'
 export { default as SisdaiColapsableNavegacion } from './colapsable-navegacion'
 export { default as SisdaiControlAcercarAlejar } from './control-acercar-alejar'
 export { default as SisdaiControlDeslizante } from './control-deslizante'
@@ -29,5 +35,7 @@ export { default as SisdaiModal } from './modal'
 export { default as SisdaiNarrativa } from './narrativa'
 export { default as SisdaiNavegacionGobMx } from './navegacion-gob-mx'
 export { default as SisdaiNavegacionPrincipal } from './navegacion-principal'
+export { default as SisdaiPestanias } from './pestanias'
 export { default as SisdaiPiePaginaConahcyt } from './pie-pagina-conahcyt'
 export { default as SisdaiPiePaginaGobMx } from './pie-pagina-gob-mx'
+export { default as SisdaiSelector } from './selector'