+++ src/componentes/pie-pagina-conahcyt/SisdaiPiePaginaConahcyt.vue
@@ -15,6 +15,7 @@
 
 <script setup>
 const anioActual = new Date().getFullYear()
+const cdn=import.meta.env.VITE_CDN_ARCHIVOS
 </script>
 
 <template>
@@ -25,7 +26,8 @@ const anioActual = new Date().getFullYear()
     <div class="flex flex-wrap">
       <div class="columna-8 columna-8-esc">
         <img
-          src="https://cdn.conacyt.mx/sisdai/recursos/conahcyt-blanco.svg"
+          class="pie-logo"
+          :src="`${cdn}navegacion/conahcyt-blanco.svg`"
           alt="Conahcyt"
           width="200px"
           height="60px"