+++ src/index.js
@@ -14,9 +14,15 @@
 //   with sisdai-componentes. If not, see <https://www.gnu.org/licenses/>.
 
 import {
+  SisdaiAreaTexto,
   SisdaiAudio,
   SisdaiBotonFlotante,
+  SisdaiBotonRadio,
+  SisdaiBotonesRadioGrupo,
+  SisdaiCampoBase,
   SisdaiCampoBusqueda,
+  SisdaiCasillaVerificacion,
+  SisdaiCasillasVerificacionGrupo,
   SisdaiColapsableNavegacion,
   SisdaiControlAcercarAlejar,
   SisdaiControlDeslizante,
@@ -30,15 +36,17 @@ import {
   SisdaiNarrativa,
   SisdaiNavegacionGobMx,
   SisdaiNavegacionPrincipal,
+  SisdaiPestanias,
   SisdaiPiePaginaConahcyt,
   SisdaiPiePaginaGobMx,
+  SisdaiSelector,
 } from './componentes'
 
 import { GlobosInformacion } from './directivas'
 
 const plugin = {
   install: function (Vue) {
-    // //UI base
+    // UI base
     Vue.use(SisdaiNavegacionGobMx)
     Vue.use(SisdaiNavegacionPrincipal)
     Vue.use(SisdaiPiePaginaGobMx)
@@ -49,29 +57,44 @@ const plugin = {
     Vue.use(SisdaiInfoDeDespliegue)
     Vue.use(SisdaiEnlaceExterno)
     Vue.use(SisdaiAudio)
-    Vue.use(SisdaiControlDeslizante)
 
-    // //accesibilidad
+    // formularios
+    Vue.use(SisdaiAreaTexto)
+    Vue.use(SisdaiBotonRadio)
+    Vue.use(SisdaiBotonesRadioGrupo)
+    Vue.use(SisdaiCampoBase)
+    Vue.use(SisdaiCampoBusqueda)
+    Vue.use(SisdaiCasillaVerificacion)
+    Vue.use(SisdaiCasillasVerificacionGrupo)
+    Vue.use(SisdaiControlAcercarAlejar)
+    Vue.use(SisdaiControlDeslizante)
+    Vue.use(SisdaiSelector)
+
+    // accesibilidad
     Vue.use(SisdaiMenuAccesibilidad)
     Vue.use(SisdaiImagenPngWebp)
 
-    // //auxiliares
+    // auxiliares
     Vue.use(GlobosInformacion)
-    // //otros
+    // otros
     Vue.use(SisdaiNarrativa)
-
     Vue.use(SisdaiBotonFlotante)
     Vue.use(SisdaiModal)
-    Vue.use(SisdaiCampoBusqueda)
-    Vue.use(SisdaiControlAcercarAlejar)
+    Vue.use(SisdaiPestanias)
   },
 }
 
 export {
+  SisdaiAreaTexto,
   SisdaiAudio,
   GlobosInformacion,
   SisdaiBotonFlotante,
+  SisdaiBotonRadio,
+  SisdaiBotonesRadioGrupo,
+  SisdaiCampoBase,
   SisdaiCampoBusqueda,
+  SisdaiCasillaVerificacion,
+  SisdaiCasillasVerificacionGrupo,
   SisdaiColapsableNavegacion,
   SisdaiControlAcercarAlejar,
   SisdaiControlDeslizante,
@@ -85,8 +108,10 @@ export {
   SisdaiNarrativa,
   SisdaiNavegacionGobMx,
   SisdaiNavegacionPrincipal,
+  SisdaiPestanias,
   SisdaiPiePaginaConahcyt,
   SisdaiPiePaginaGobMx,
+  SisdaiSelector,
 }
 
 export default plugin