+++ CHANGELOG.md
@@ -11,9 +11,92 @@ de la biblioteca. Por lo tanto no se listan aquí todos los tags existentes.
 Al principio de cada entrada se lista la versión de la biblioteca de sisdai-css con la
 que la versión indicada de sisdai-componentes es compatible y tiene instalada.
 
+## [4.12.0] - 2024-12-16
+
+### Agregado (Added)
+
+- Se agrega el componente SisdaiPestanias
+
+## [4.11.0] - 2024-12-16
+
+### Agregado (Added)
+
+- Se agregaron metadatos de información para compartir la página de documentación.
+- Se agregaron variables de entorno.
+
+### Fixed (Changed)
+
+- Se ajustó la orientación de las tarjetas del inicio.
+- Se ajustaron rutas de imágenes.
+
+## [4.10.2] - 2024-12-12
+
+### Arreglado (Fixed)
+
+- Se recupera la documentación del vitepress en las secciones de Instalación y Uso de componentes
+- Se arregla la propiedad indefinida item en el Ìndice de contenido en el Layout
+
+## [4.10.1] - 2024-12-13
+
+### Arreglado (Fixed)
+
+- Se arreglo la ruta hacia los componentes area-texto, boton-radio, campo-base, casilla-verificacion y selector
+
+## [4.10.0] - 2024-12-12
+
+### Cambiado (Changed)
+
+- Se actualiza sisdai-css v1.7.2
+
+## [4.9.0] - 2024-12-12
+
+### Cambiado (Changed)
+
+- Se actualiza maqueta de documentación
+
+## [4.8.0] - 2024-11-29
+
+### Agregado (Added)
+
+- Se agrega el componente SisdaiBotonRadio y SisdaiBotonesRadioGrupo para mejorar la accesibilidad en elementos de botones de radio para formularios
+
+## [4.7.0] - 2024-11-29
+
+### Agregado (Added)
+
+- Se agrega el componente SisdaiCasillaVerificacion y SisdaiCasillasVerificacionGrupo para mejorar la accesibilidad en elementos de casillas de verificación
+
+## [4.6.1] - 2024-11-29
+
+sisdai-css v1.5.0
+
+### Cambiado (Changed)
+
+- Cambiamos algunas partes de la documentación para especificar que la importación debe ser de `@centrogeomx/sisdai-componentes` en lugar de `sisdai-componentes`.
+- Se agregó la clase css 'pie-logo' a las imágenes de los pies para que evitar que se oculten al activar Mostrar solo texto en el menú de accesibilidad.
+- Se actualizó la versión de `sisdai-css` a la v1.5.0
+
+## [4.6.0] - 2024-11-25
+
+### Agregado (Added)
+
+- Se agrega el componente SisdaiSelector para mejorar uso del elemento selector del formulario
+
+## [4.5.0] - 2024-11-25
+
+### Agregado (Added)
+
+- Se agrega el componente SisdaiAreaTexto para uso del elemento de campo de texto del formulario
+
+## [4.4.0] - 2024-11-25
+
+### Agregado (Added)
+
+- Se agrega el componente SisdaiCampoBase para uso general de la mayoría de los campos del formulario
+
 ## [4.3.2] - 2024-11-21
 
-Arreglado (Fixed)
+## Arreglado (Fixed)
 
 - Url del repositorio de npm de la biblioteca en el README.md
 
@@ -53,7 +136,6 @@ Arreglado (Fixed)
 - Script y archivos de variable de entorno para seguimiento de analíticas en Matomo
 
 ## [4.1.6] - 2024-11-04
-
 
 ### Agregado (Added)
 
@@ -128,8 +210,8 @@ para el periodo 2024-2030
 
 sisdai-css v1.3.1. ó v1.3.2
 
-
 ### Cambiado (Changed)
+
 - Corrección de errores de Sass
 
 ## [3.1.5] - 2024-09-23
@@ -138,7 +220,6 @@ Versiones deprecadas de Vue 2 con la identidad gráfica del gobierno federal
 para el periodo 2018-2024
 
 sisdai-css v1.2.7
-
 
 ## [3.0.0] - 2024-07-08 - Release
 
@@ -166,19 +247,29 @@ Segunda versión estable
 
 sisdai-css v0.37.3
 
-
 Primera versión estable
-
-[4.3.2]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.3.2...v4.3.1
-[4.3.1]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.3.1...v4.2.1
-[4.2.1]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.2.1...v4.2.0
-[4.2.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.2.0...v4.1.6
-[4.1.6]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.1.6...v4.1.5
-[4.1.5]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.1.5...v4.1.0
-[4.1.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.1.0...v4.0.0
-[4.0.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.1.0...v3.2.1
-[3.2.1]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v3.2.1...v3.1.5
-[3.1.5]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v3.1.5...v3.0.0
-[3.0.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v3.0.0...v2.0.0
-[2.0.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v2.0.0...v1.0.0
+[4.12.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.11.0...v4.12.0
+[4.11.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.10.2...v4.11.0
+[4.10.2]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.10.1...v4.10.2
+[4.10.1]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.10.0...v4.10.1
+[4.10.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.9.0...v4.10.0
+[4.9.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.8.0...v4.9.0
+[4.8.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.7.0...v4.8.0
+[4.7.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.6.1...v4.7.0
+[4.6.1]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.6.0...v4.6.1
+[4.6.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.5.0...v4.6.0
+[4.5.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.4.0...v4.5.0
+[4.4.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.3.2...v4.4.0
+[4.3.2]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.3.1...v4.3.2
+[4.3.1]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.2.1...v4.3.1
+[4.2.1]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.2.0...v4.2.1
+[4.2.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.1.6...v4.2.0
+[4.1.6]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.1.5...v4.1.6
+[4.1.5]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.1.0...v4.1.5
+[4.1.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.0.0...v4.1.0
+[4.0.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v4.0.0...v3.2.1
+[3.2.1]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v3.1.5...v3.2.1
+[3.1.5]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v3.0.0...v3.1.5
+[3.0.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v2.0.0...v3.0.0
+[2.0.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/compare/v1.0.0...v2.0.0
 [1.0.0]: https://codigo.conahcyt.mx/sisdai/sisdai-componentes/-/releases/v1.0.0